+++ web-frontend/modules/database/components/view/grid/fields/GridViewFieldLinkRow.vue
@@ -89,7 +89,7 @@ export default {
       // While the field is selected we want to open the select row popup by pressing
       // the enter key.
       this.$el.keydownEvent = (event) => {
-        if (event.keyCode === 13 && !this.modalOpen) {
+        if (event.key === 'Enter' && !this.modalOpen) {
           this.showModal()
         }
       }