+++ web-frontend/modules/database/components/view/grid/fields/GridViewFieldLongText.vue
@@ -26,7 +26,7 @@ export default {
   methods: {
     afterEdit(event) {
       // If the enter key is pressed we do not want to add a new line to the textarea.
-      if (event.type === 'keydown' && event.keyCode === 13) {
+      if (event.type === 'keydown' && event.key === 'Enter') {
         event.preventDefault()
       }
       this.$nextTick(() => {
@@ -34,8 +34,9 @@ export default {
         this.$refs.input.selectionStart = this.$refs.input.selectionEnd = 100000
       })
     },
-    canSaveByPressingEnter() {
-      return false
+    canSaveByPressingEnter(event) {
+      // Save only if shiftKey is pressed
+      return event.shiftKey
     },
   },
 }