+++ web-frontend/modules/database/mixins/bufferedRowsDragAndDrop.js
@@ -82,7 +82,7 @@ export default {
       this.clonedWrapper.appendChild(this.clonedElement)
 
       this.$el.keydownEvent = (event) => {
-        if (event.keyCode === 27) {
+        if (event.key === 'Escape') {
           if (this.dragAndDropDraggingRow !== null) {
             this.$store.dispatch(
               `${this.getDragAndDropStoreName(this)}/cancelRowDrag`,